Add working test case for Transaction save signal
The post_save signal emitted by django is connected to the method to
update the account balances of the payer and the payee involved in the
transaction.
This modifiction itself is done using F expression in django, which is
supposed to be more robuts that directly assigning a new value for the
field and calling save on the model for writing the changes to database.

A big assumption here is that the transaction added once can not be
removed, as there is no implementation for the update of the transaction
as of now. There are some possibilities of hooking up this with the
pre_save signal, and fetching the old value of the balance from the db,
and then adjusting the accounts' balance accordingly, but that requires
custom code in the model afaik.